feat: add errors to PageProps
This commit adds an `errors` property to the `PageProps` type definition in `resources/js/types/index.d.ts`. This property is a `Record<string, string[]>` and will be used to pass validation errors from the server to the client-side components. This allows for displaying form validation errors to the user.