feat: add Head component and meta tags to Login page
This commit adds the `Head` component from `@inertiajs/vue3` to the `Login.vue` page. This allows for setting the page title and meta tags. Specifically, it sets the title to "Login" and adds a meta description. This improves SEO and provides better information to search engines and social media platforms.